skintel-facial-landmarks/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
python-jose[cryptography]==3.3.0
pydantic==2.5.0
requests==2.31.0
# dlib - install separately, see install_instructions.md
dlib
